% Generated by roxygen2 (4.0.2): do not edit by hand
\name{alterData}
\alias{alterData}
\title{A function which takes the input data from a table in the interface and returns the new data for that table}
\usage{
alterData(...)
}
\description{
A function which takes the input data from a table in the interface and returns the new data for that table
}

