% Generated by roxygen2 (4.0.2): do not edit by hand
\name{baseData}
\alias{baseData}
\title{Generate Base Demo Data}
\usage{
baseData(pickups = 1000, numTrucks = 100, numStations = 10,
  numDropOffs = 30)
}
\arguments{
\item{pickups}{is the number of pickups to select from the data set}
}
\value{
data.frame with the wells, pickups and etc.
}
\description{
Generate Base Demo Data
}

