% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dispatchBriefSummary}
\alias{dispatchBriefSummary}
\title{A function which calculates summary metrics on the dispatch queue}
\usage{
dispatchBriefSummary(dispatchSummary, baseData)
}
\arguments{
\item{dispatch}{a list object containing the dispatch queue}
}
\value{
list with a handful of metric summaries
}
\description{
A function which calculates summary metrics on the dispatch queue
}

