% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dispatchQueue}
\alias{dispatchQueue}
\title{Dispatch a set of scheduled pickups}
\usage{
dispatchQueue(data, pickupWindow = 1, windowsAhead = 1)
}
\arguments{
\item{data}{a list of various data items returned from baseData, or passed in via JSON}
}
\value{
list with a data.frame of dispatches
}
\description{
Dispatch a set of scheduled pickups
}

