% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dispatchSummaries}
\alias{dispatchSummaries}
\title{A function which calculates summary metrics on the dispatch queue}
\usage{
dispatchSummaries(dispatch, baseData)
}
\arguments{
\item{dispatch}{a list object containing the dispatch queue}
}
\value{
list with several data.frames containing metric summaries
}
\description{
A function which calculates summary metrics on the dispatch queue
}

