% Generated by roxygen2 (4.0.2): do not edit by hand
\name{great_circle_distance}
\alias{great_circle_distance}
\title{Calculate distances over the curvature of the earth}
\usage{
great_circle_distance(lat1, long1, lat2, long2)
}
\arguments{
\item{lat1}{a vector of lattitudes to start at}

\item{long1}{a vector of longitudes to start at}

\item{lat2}{a vector of lattitues to end at}

\item{long2}{a vector of longitudes to end at}
}
\value{
numeric vector of distances
}
\description{
Calculate distances over the curvature of the earth
}

