% Generated by roxygen2 (4.0.2): do not edit by hand
\name{optimizeBaseStations}
\alias{optimizeBaseStations}
\title{Take the generated Base Stations and figure out where they should move to to be better placed}
\usage{
optimizeBaseStations(baseData)
}
\arguments{
\item{baseData}{is the list object returned from the function baseData()}
}
\value{
data.frame with the current baseStations, their position, the recommended position and the distance between the two, and average distance reduction between it and the surrounding pickups, dropOffs and their weights
}
\description{
Take the generated Base Stations and figure out where they should move to to be better placed
}

