% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pickupDropOffs}
\alias{pickupDropOffs}
\title{A function which helps create a pickupDropOffPair Matrix}
\usage{
pickupDropOffs(pickupsWin, dropOffs, baseStations)
}
\arguments{
\item{pickups}{a dataframe containing pickups to be made}

\item{baseData}{the list object returned from baseData()}
}
\description{
A function which helps create a pickupDropOffPair Matrix
}

