% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pickupOpt}
\alias{pickupOpt}
\title{A function which does the actual dispatching and metric calculations}
\usage{
pickupOpt(pickupsWin, baseStations, trucksStations, trucksInfo, dropOffs, ...,
  truckLocation = NULL)
}
\arguments{
\item{pickupsWin}{}

\item{baseStations}{}

\item{trucksStations}{}

\item{trucksInfo}{}

\item{dropOffs}{}

\item{truckLocation}{}
}
\description{
A function which does the actual dispatching and metric calculations
}

