% Generated by roxygen2 (4.0.2): do not edit by hand
\name{returnTable}
\alias{returnTable}
\title{This function takes a list (like baseData) and a table name and returns just that table}
\usage{
returnTable(data, table)
}
\description{
This function takes a list (like baseData) and a table name and returns just that table
}

